+++ Changelog.md
@@ -10,6 +10,7 @@
 ### Breaking changes
 
 - The standard function `ToString()` now depends on the active culture (`CultureInfo.CurrentCulture`). The reason is that NaN and Infinity should be displayed in the system language or the corresponding symbol should be used.
+- Argument name for `Fraction.TryParse(..)` changed from `fractionString` to `value`.
 - A fraction of 0/0 no longer has the value 0, but means NaN (not a number). Any fraction in the form x/0 is no longer a valid number. A denominator with the value 0 corresponds to, depending on the numerator, NaN, PositiveInfinity or NegativeInfinity.
 
 ## 7.7.1